QGA
Quantum Genetic Algorithm (QGA) is a QGA that can be used for the purposes of education and research. QGA is applied in a simple optimization problem: Let  f(x)=abs(x-5/2+sin(x)) be a function that takes values in the range 0<=x<=15. Within this range f(x) has a maximum value at x=11 (binary value is equal to 1011).