HGA
Hybrid Genetic Algorithm (HGA) is a GA that combines quantum operators (rotation, measure, quantum chromosomes, etc.) with classical genetic operators (crossover and mutation).It  can be used for the purposes of education and research. HGA is applied in a simple optimization problem: Let  f(x)=abs(x-5/2+sin(x)) be a function that takes values in the range 0<=x<=15. Within this range f(x) has a maximum value at x=11 (binary value is equal to 1011).