RQGA
Reduced Quantum Genetic Algorithm (RQGA) is a program in Python showing how to implement a 'true' quantum genetic algorithm based on a fitness quantum gate and Grover's search algorithm. It  can be used for the purposes of education and research. RQGA is applied in a simple optimization problem: Let  f(x)=abs(x-5/2+sin(x)) be a function that takes values in the range 0<=x<=15. Within this range f(x) has a maximum value at x=11 (binary value is equal to 1011).